+++ src/frontend/src/components/DataConflation/SubmissionConflation/MergeAttributes.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { Modal } from '@/components/common/Modal';
-import Button from '@/components/common/Button2';
+import Button from '@/components/common/Button';
 import Table, { TableHeader } from '@/components/common/CustomTable';
 
 type mergeAttributesPropType = {
