+++ src/frontend/src/components/ManageProject/DeleteTab/ProjectDeleteTab.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Button from '@/components/common/Button2';
+import Button from '@/components/common/Button';
 import { Modal } from '@/components/common/Modal';
 import { DeleteProjectService } from '@/api/CreateProjectService';
 import { useAppDispatch } from '@/types/reduxTypes';
