+++ src/frontend/src/components/ProjectSubmissions/TaskSubmissions.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import TaskSubmissionsMap from '@/components/ProjectSubmissions/TaskSubmissionsMap';
 import InputTextField from '@/components/common/InputTextField';
-import Button from '@/components/common/Button2';
+import Button from '@/components/common/Button';
 import AssetModules from '@/shared/AssetModules.js';
 import CoreModules from '@/shared/CoreModules.js';
 import { TaskCardSkeletonLoader } from '@/components/ProjectSubmissions/ProjectSubmissionsSkeletonLoader';
