+++ src/frontend/src/components/createnewproject/ProjectDetailsForm.tsx
@@ -6,7 +6,7 @@ import { useNavigate } from 'react-router-dom';
 import { useAppDispatch, useAppSelector } from '@/types/reduxTypes';
 import useForm from '@/hooks/useForm';
 import CreateProjectValidation from '@/components/createnewproject/validation/CreateProjectValidation';
-import Button from '@/components/common/Button2';
+import Button from '@/components/common/Button';
 import { CommonActions } from '@/store/slices/CommonSlice';
 import { OrganisationService } from '@/api/CreateProjectService';
 import { CustomCheckbox } from '@/components/common/Checkbox';
