+++ src/frontend/src/components/createnewproject/SplitTasks.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import Button from '@/components/common/Button2';
+import Button from '@/components/common/Button';
 import RadioButton from '@/components/common/RadioButton';
 import AssetModules from '@/shared/AssetModules.js';
 import { CommonActions } from '@/store/slices/CommonSlice';
